api/requirements.txt
sqlmodel
fastapi[standard]
pytest
pyjwt
pydantic-settings
email-validator
argon2-cffi